macOS:  Use available keyword instead of preprocessor version checks where possible

+++ modules/juce_core/native/juce_mac_Files.mm
@@ -285,13 +285,17 @@ bool File::moveToTrash() const
 
     JUCE_AUTORELEASEPOOL
     {
-       #if (defined (__IPHONE_11_0) && __IPHONE_OS_VERSION_MIN_REQUIRED >= __IPHONE_11_0) \
-         || (defined (MAC_OS_X_VERSION_10_8) && MAC_OS_X_VERSION_MIN_REQUIRED >= MAC_OS_X_VERSION_10_8)
-        NSError* error = nil;
-        return [[NSFileManager defaultManager] trashItemAtURL: createNSURLFromFile (*this)
-                                             resultingItemURL: nil
-                                                        error: &error];
-       #elif JUCE_IOS
+       #if JUCE_MAC || (JUCE_IOS && (defined (__IPHONE_11_0) && __IPHONE_OS_VERSION_MAX_ALLOWED >= __IPHONE_11_0))
+        if (@available (macOS 10.8, iOS 11.0, *))
+        {
+            NSError* error = nil;
+            return [[NSFileManager defaultManager] trashItemAtURL: createNSURLFromFile (*this)
+                                                 resultingItemURL: nil
+                                                            error: &error];
+        }
+       #endif
+
+       #if JUCE_IOS
         return deleteFile();
        #else
         [[NSWorkspace sharedWorkspace] recycleURLs: [NSArray arrayWithObject: createNSURLFromFile (*this)]
